[adapted]simple_lmk: Reduce unnecessary wake ups
We can check if the waitqueue is actually active before calling wake_up()
in order to avoid an unnecessary wake_up() if the reclaim thread is already
running. Furthermore, the release barrier when zeroing needs_reclaim is
unnecessary, so remove it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

pwnrazr: adapted for 41c5c9760dedf488acfc271fac702930367619e8